Work has begun on the content-main-head and content-main-services blocks.
html - Added tags, elements and classes for content-main-services and content-main-head blocks, added mixes to some tags.

css - added the main classes of the content-main-head and content-main-services blocks to the style.css file. Added basic indentation and flex elements.

assets - added images to the folder img/content-main to work with blocks.

@AidenOrtiz